+++ src/aks-preview/azext_aks_preview/tests/latest/test_run_command.py
@@ -74,6 +74,5 @@ class TestRunCommand(ScenarioTest):
             self.check('exitCode', 0),
         ])
 
-
 if __name__ == '__main__':
     unittest.main()