+++ src/spring-cloud/azext_spring_cloud/_client_factory.py
@@ -5,7 +5,12 @@
 from azure.cli.core.commands.client_factory import get_mgmt_service_client
 from azure.cli.core.profiles import ResourceType
 from .vendored_sdks.appplatform.v2020_07_01 import AppPlatformManagementClient
-from .vendored_sdks.appplatform.v2020_11_01_preview import AppPlatformManagementClient as AppPlatformManagementClient_20201101preview
+from .vendored_sdks.appplatform.v2020_11_01_preview import (
+    AppPlatformManagementClient as AppPlatformManagementClient_20201101preview
+)
+from .vendored_sdks.appplatform.v2021_06_01_preview import (
+    AppPlatformManagementClient as AppPlatformManagementClient_20210601preview
+)
 
 
 def cf_spring_cloud(cli_ctx, *_):
@@ -14,6 +19,10 @@ def cf_spring_cloud(cli_ctx, *_):
 
 def cf_spring_cloud_20201101preview(cli_ctx, *_):
     return get_mgmt_service_client(cli_ctx, AppPlatformManagementClient_20201101preview)
+
+
+def cf_spring_cloud_20210601preview(cli_ctx, *_):
+    return get_mgmt_service_client(cli_ctx, AppPlatformManagementClient_20210601preview)
 
 
 def cf_resource_groups(cli_ctx, subscription_id=None):