+++ src/spring-cloud/azext_spring_cloud/vendored_sdks/appplatform/v2019_05_01_preview/models/__init__.py
@@ -46,6 +46,7 @@ try:
     from ._models_py3 import PersistentDisk
     from ._models_py3 import ProxyResource
     from ._models_py3 import RegenerateTestKeyRequestPayload
+    from ._models_py3 import RequiredTraffic
     from ._models_py3 import Resource
     from ._models_py3 import ResourceSku
     from ._models_py3 import ResourceSkuCapabilities
@@ -101,6 +102,7 @@ except (SyntaxError, ImportError):
     from ._models import PersistentDisk
     from ._models import ProxyResource
     from ._models import RegenerateTestKeyRequestPayload
+    from ._models import RequiredTraffic
     from ._models import Resource
     from ._models import ResourceSku
     from ._models import ResourceSkuCapabilities
@@ -131,6 +133,7 @@ from ._app_platform_management_client_enums import (
     ProvisioningState,
     ConfigServerState,
     TraceProxyState,
+    TrafficDirection,
     ManagedIdentityType,
     TestKeyType,
     AppResourceProvisioningState,
@@ -182,6 +185,7 @@ __all__ = [
     'PersistentDisk',
     'ProxyResource',
     'RegenerateTestKeyRequestPayload',
+    'RequiredTraffic',
     'Resource',
     'ResourceSku',
     'ResourceSkuCapabilities',
@@ -211,6 +215,7 @@ __all__ = [
     'ProvisioningState',
     'ConfigServerState',
     'TraceProxyState',
+    'TrafficDirection',
     'ManagedIdentityType',
     'TestKeyType',
     'AppResourceProvisioningState',